TPU debugger
TDB (Tensor debugger/ TPU debugger) is a tool to debug the BModel running on
TPU. It provides tracking BModel execution and monitor changes in TPU resources
that may indicate malfunctioning code. TDB should include the ability to run or
halt the target program at specific points, display the contents of global
memory, SRAM, or L2 memory, and modify memory or register contents to enter
selected test data that might be a cause of faulty program execution.

- Current feature:
  1. Running a BModel step by step
  2. Filling inputs data and getting outputs data
  3. Showing the disassembled code of BModel
  4. Getting any data used or produced by an instruction.

- TODO:
  1. bugfix
  2. breakpoint
  3. step backward
  4. dump data
  5. data monitor
  6. print code and location

Change-Id: I9fba3a9bfe6e3d0672a51d03ef788410306776de